Make integration tests storage agnostic
Additional storage engines need to be added so the tests can't depend on
mongo.
